// frontend/src/app/core/services/api/model/transferRounding.ts
/**
 * BFG API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Rounding strategy for crew transfer calculations
 */
export const TransferRounding = {
    Floor: 'FLOOR',
    Ceil: 'CEIL',
    Round: 'ROUND'
} as const;
export type TransferRounding = typeof TransferRounding[keyof typeof TransferRounding];